added function to create a colored fiber density texture

+++ src/gui/gl/arcball.h
@@ -26,7 +26,7 @@ public:
     void midClick( int x, int y );
     void midDrag( int x, int y );
 
-    void mouseWheel( int step );
+    void mouseWheel( float step );
 
     void setRotCenter( float x, float y, float z );
